% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Runs one "monte carlo" simulation trial of the ABM.}
\usage{
sim(sim.params)
}
\arguments{
\item{sim.params}{List of options governing simulation size and behavior:
Value ranges are marked as [lower, upper)\link{default}.
n.candidates: Number of alternatives/candidates in the field [1,inf)\link{10}
n.dims: Maximum number of political issues affecting utility [1, 32)\link{5}
n.agents: Number of agents to simulate [1, inf)\link{200}
n.heuristics: Number of distinct combinations of issues allowed \link[=32]{1, 2^n_dims}
n.timesteps: Number of ticks to run the simulation [1, inf)\link{250}
election.cycle: Ticks between elections. 0 cancels elections. \link[=0]{0, n_timesteps}
conservatism: Agent motivation to preserve current way of thinking \link[=0]{0, 1}
ego: Agent selfishness vs selflessness in utility calculation \link[=1]{0, 1}
aversion: Agent aversion to complexity [0, 5)\link{1}
org.aversion: Political org aversion to complexity (0, 5)\link{1}}
}
\value{
A list of many results matrixes
}
\description{
Runs one "monte carlo" simulation trial of the ABM.
}
